refactor: 애니메이션 컴포넌트 분리
